refactor(clustering): move the clustering result types into their own package
Second of the stack. Types only — no algorithm moves here.

ClusterResult and METHOD_LEVEL_STRATEGY describe the *output* of clustering, and
MethodClusterPaths tracks each method's scoped lineage; none of the three is
graph structure. They become static_analyzer/clustering, which graph.py now
imports from rather than declares.

The Leiden search itself stays on CallGraph for now and moves in the next PR,
so this one is a pure relocation of type declarations plus the import updates
that follow from it.
