+++ agents/incremental_results.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass, field
 
-from static_analyzer.graph import CallGraph, ClusterResult
+from static_analyzer.graph import CallGraph
+from static_analyzer.clustering import ClusterResult
 
 
 @dataclass(frozen=True)
