+++ tests/agents/test_cluster_methods_mixin.py
@@ -12,7 +12,8 @@ from agents.agent_responses import (
     SourceCodeReference,
 )
 from agents.file_index_models import FileMethodGroup, MethodEntry
-from static_analyzer.graph import CallGraph, ClusterResult
+from static_analyzer.graph import CallGraph
+from static_analyzer.clustering import ClusterResult
 from static_analyzer.constants import Language, NodeType
 from static_analyzer.node import Node
