+++ tests/agents/test_incremental_agent.py
@@ -27,7 +27,8 @@ from agents.incremental_agent import (
 from agents.incremental_results import ScopeRelationContext
 from static_analyzer.analysis_result import StaticAnalysisResults
 from static_analyzer.constants import NodeType
-from static_analyzer.graph import CallGraph, ClusterResult
+from static_analyzer.graph import CallGraph
+from static_analyzer.clustering import ClusterResult
 from static_analyzer.node import Node
