+++ tests/diagram_analysis/test_tree_shape.py
@@ -11,7 +11,7 @@ from diagram_analysis.analysis_json import build_unified_analysis_json, parse_un
 from diagram_analysis.diagram_generator import DiagramGenerator, _member_keys, _reconcile_child_scope
 from diagram_analysis.exceptions import ScopeContainmentError
 from static_analyzer.graph import CallGraph
-from static_analyzer.method_cluster_paths import MethodClusterPaths
+from static_analyzer.clustering.method_cluster_paths import MethodClusterPaths
 from diagram_analysis.tree_shape import absorb_single_child_components
